requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2-binary==2.9.1
python-decouple==3.4asgiref==3.4.1
Django==3.2.7
djangorestframework==3.
Pillow==8.3.2
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
django-filter
